13234: add sessionsThunks
